+++ .gitignore
@@ -13,3 +13,7 @@ src/main/resources/docs/
 *.iml
 bin/
 ACTUAL.TXT
+DateTime.class
+Deadline.class
+Duke.class
+*.class
